scripts/carray: change to using single .awk script to generate .c
The shell script makes multiple call-outs to awk to get information
from the configuration file. It would be easier to just write the
whole thing in one .awk script and have the makefile altered to call
that instead.

There should be no functional difference other than the script type
and has been tested with PLATFORM=generic build. This should be both
quicker and easier to understand.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>
---
v2:
  - updated with the new bits from Samuel Holland's changes
  - minor tweaks to the awk script to fix the output spacings